CM_AUTO_GROUP/AutoGroupService byte<->sbyte: (sbyte) cast entryRequestId to GetTypeById + ert decl EntryRequestType? + ert.Value to StartLooking + (byte) cast agt.GetDifficultId() to GetNextAvailableInstance (faithful Java byte=sbyte, GetTypeById nullable); clears 2 CS1503 (889->887)

+++ dotnetConversion/src/Aion.GameServer/Services/AutoGroupService.cs
@@ -106,7 +106,7 @@ public class AutoGroupService
 
     private void CreateNewInstance(AutoInstance autoInstance, AutoGroupType agt, List<LookingForParty> filteredParties, int maskId)
     {
-        WorldMapInstance instance = InstanceService.GetNextAvailableInstance(agt.GetTemplate().GetInstanceMapId(), 0, agt.GetDifficultId(), null,
+        WorldMapInstance instance = InstanceService.GetNextAvailableInstance(agt.GetTemplate().GetInstanceMapId(), 0, (byte)agt.GetDifficultId(), null,
             autoInstance.GetMaxPlayers(), false);
         autoInstance.OnInstanceCreate(instance);
         autoInstances[instance] = autoInstance;
